requirement.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
aiofiles
boto3
google-cloud-vision
pillow
opencv-python-headless
pytesseract
openai
pdf2image
pytest
requests
PyMuPDF
